Hack to fix the following issue:
Sort of - source is string value and the code should have exported it as
'' rather than a blank string. I'll take a look. In the meantime you
could put the eval statement in a try except block something like:

try:
    ret=eval(....)
except SyntaxError:
   ret=""

which should at least stop the crash and burn...