Fix of custom header for error column
When using basic math operations on columns, uncertainty columns can be created. However creation of the custom header for such column was not possible. In the first commit of these functions, there was a typo len(header) == 0, later changed to "not header". But neither worked, len of the tupple needs to be two, as for a and b variables on line 2530.